Function for downloading image from URL/ Prototype for downloading the image and doing silhouette at the same time.
I researched methods for using requests to download image URL's and created a function to do so. Additionally, I researched how to download the image and open as a pillow image for manipulation and created a function prototype for integration with silhouette. At this junction I see two different ways we could go about accomplishing this task and would like to discuss the best way to do this with the group!